+++ addon/utils/_private/ember-node.ts
@@ -1,3 +1,17 @@
+/**
+ * Contains code from https://github.com/ueberdosis/tiptap/blob/d61a621186470ce286e2cecf8206837a1eec7338/packages/core/src/NodeView.ts#L195
+ *
+ * MIT License
+ * Copyright (c) 2023, Tiptap GmbH
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
+ * The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+
+/**
+
+ */
+
 import { hbs, TemplateFactory } from 'ember-cli-htmlbars';
 import {
   AttributeSpec,
@@ -26,7 +40,7 @@ import { SayView } from '@lblod/ember-rdfa-editor';
  * - Prosemirror NodeView e.g.:
  *   - `ignoreMutation`: Use this to avoid rerendering a component for every change.
  *        Already as a default implementation. Only override if you know what you are doing.
- *   - `stopEvent`: defaults to true (stopping all event bubbling). Implement this if you need events to bubble up.
+ *   - `stopEvent`: By default this will stop events which occur inside the ember-node but not inside it's content. Only override if you know what you are doing.
  * - Custom values for EmberNode, e.g.:
  *   - `componentPath`: path to the ember component to render as a Node View
  *
@@ -41,6 +55,33 @@ import { SayView } from '@lblod/ember-rdfa-editor';
  *       Instead of a tracked property, you'll often use the following logic to keep state inside the node:
  *       `get someText() { return this.args.node.attrs.someText; }`
  *       `set someText(value) { return this.args.updateAttribute('someText', value); }`
+ *
+ *   - when defining an ember-node using these utility functions, nested contenteditable attributes should be prevented.
+ *     Ember nodes which can contain editable content should never be `contenteditable: false` as a whole,
+ *     but only the parts which are `contenteditable: false` should be marked as so.
+ *     E.g.: It's preferred to write:
+ *
+ *     ```
+ *     <div>
+ *     <header contenteditable="false">
+ *      <p>header</p>
+ *     </header>
+ *     <content>
+ *      {{yield}}
+ *     </content>
+ *     </div>
+ *     ```
+ *     instead of
+ *     ```
+ *     <div contenteditable="false">
+ *     <header>
+ *      <p>header</p>
+ *     </header>
+ *     <content contenteditable="true">
+ *      {{yield}}
+ *     </content>
+ *     </div>
+ *     ```
  */
 
 export interface EmberInlineComponent extends Component, EmberNodeArgs {
@@ -94,8 +135,7 @@ class EmberNodeView implements NodeView {
   contentDOM?: HTMLElement;
   emberComponent: EmberInlineComponent;
   template: TemplateFactory;
-  stopEvent: (event: InputEvent) => boolean;
-  ignoreMutation: (mutation: MutationRecord) => boolean;
+  config: EmberNodeConfig;
 
   constructor(
     controller: SayController,
@@ -106,16 +146,8 @@ class EmberNodeView implements NodeView {
   ) {
     // when a node gets updated, `update()` is called.
     // We set the new node here and pass it to the component to render it.
-    // However, this will create a DOM mutation, which prosemirror will catch and use to rerender.
-    // to avoid a NodeView rerendering when we already handled the state change, we pass true to `ignoreMutation` (by default).
-    const {
-      name,
-      componentPath,
-      atom,
-      inline,
-      stopEvent = () => true,
-      ignoreMutation = (_) => true,
-    } = emberNodeConfig;
+    this.config = emberNodeConfig;
+    const { name, componentPath, atom, inline } = emberNodeConfig;
     this.template = hbs`{{#component this.componentPath
                           getPos=this.getPos
                           node=this.node
@@ -131,11 +163,12 @@ class EmberNodeView implements NodeView {
                         {{/component}}`;
     this.node = pNode;
     this.contentDOM = !atom
-      ? document.createElement(inline ? 'span' : 'div')
+      ? document.createElement(inline ? 'span' : 'div', {})
       : undefined;
-    this.stopEvent = stopEvent;
-    this.ignoreMutation = ignoreMutation;
-
+    // Note `this.contentDOM` needs an attribute to prevent chromium-based browsers from deleting it when it is empty/only has empty children.
+    if (this.contentDOM) {
+      this.contentDOM.dataset.emberNodeContent = 'true';
+    }
     const { node, component } = emberComponent(
       controller.owner,
       name,
@@ -189,6 +222,72 @@ class EmberNodeView implements NodeView {
   destroy() {
     this.emberComponent.destroy();
   }
+
+  /**
+   *
+   * Prevents the editor view from handling events which are inside the ember-node but not inside it's editable content.
+   * Based on https://github.com/ueberdosis/tiptap/blob/d61a621186470ce286e2cecf8206837a1eec7338/packages/core/src/NodeView.ts#LL99C6-L99C6
+   * @param event The event to check
+   */
+  stopEvent(event: Event) {
+    if (!this.dom) {
+      return false;
+    }
+
+    if (this.config.stopEvent) {
+      return this.config.stopEvent(event);
+    }
+
+    const target = event.target as HTMLElement;
+    const isInElement =
+      this.dom.contains(target) && !this.contentDOM?.contains(target);
+
+    return isInElement;
+  }
+
+  /**
+   *
+   * Determines whether a DOM mutation should be ignored by prosemirror.
+   * DOM mutations occuring inside the ember-node which are not inside it's editable content are ignored.
+   * Selections are always handled by prosemirror.
+   * Taken from https://github.com/ueberdosis/tiptap/blob/d61a621186470ce286e2cecf8206837a1eec7338/packages/core/src/NodeView.ts#L195
+   * @param mutation
+   * @returns
+   */
+  ignoreMutation(
+    mutation: MutationRecord | { type: 'selection'; target: Element }
+  ) {
+    if (!this.dom || !this.contentDOM) {
+      return true;
+    }
+    if (this.config.ignoreMutation) {
+      return this.config.ignoreMutation(mutation);
+    }
+
+    // a leaf/atom node is like a black box for ProseMirror
+    // and should be fully handled by the node view
+    if (this.node.isLeaf || this.node.isAtom) {
+      return true;
+    }
+
+    // ProseMirror should handle any selections
+    if (mutation.type === 'selection') {
+      return false;
+    }
+
+    // we will allow mutation contentDOM with attributes
+    // so we can for example adding classes within our node view
+    if (this.contentDOM === mutation.target && mutation.type === 'attributes') {
+      return true;
+    }
+
+    // ProseMirror should handle any changes within contentDOM
+    if (this.contentDOM.contains(mutation.target)) {
+      return false;
+    }
+
+    return true;
+  }
 }
 
 export type EmberNodeConfig = {
@@ -210,8 +309,10 @@ export type EmberNodeConfig = {
   };
   parseDOM?: readonly ParseRule[];
   toDOM?: (node: PNode) => DOMOutputSpec;
-  stopEvent?: (event: InputEvent) => boolean;
-  ignoreMutation?: (mutation: MutationRecord) => boolean;
+  stopEvent?: (event: Event) => boolean;
+  ignoreMutation?: (
+    mutation: MutationRecord | { type: 'selection'; target: Element }
+  ) => boolean;
 } & (
   | {
       atom: true;